+++ protected/humhub/modules/user/controllers/RegistrationController.php
@@ -156,8 +156,11 @@ class RegistrationController extends Controller
             'scenario' => Invite::SCENARIO_INVITE_BY_LINK_FORM,
         ]);
         if ($form->load(Yii::$app->request->post()) && $form->validate()) {
-            $invite = $linkRegistrationService->convertToInvite($form->email);
-            $invite?->sendInviteMail();
+            if ($invite = $linkRegistrationService->convertToInvite($form->email)) {
+                $invite->sendInviteMail();
+            } else {
+                $invite = new Invite(['email' => $form->email]);
+            }
             return $this->render('@user/views/auth/register_success', ['model' => $invite]);
         }